Keys/Keychains/item functionality/ Light modifications, Locked/Unlocked door functionality
Item Functionality Basis and Locked Doors

Enhanced Light System so that all lights dont glare alongside the Master Light

Both Keys and KeyChains can open/close doors

You'll need to look at the UnityEditor as well as the code to make sure things still work properly from this point onwards. EX: A door  that leads to darkness, while the master light is on, may be a case of the door not being registered in the MasterLight Script as a variable to keep track of. We also learned the magic of  static variables (not static classes, static variables).

That whole

public class Xavier

public static Xavier X

public Start {X = this}

stuff.

saves a lot of time from having to take in variables like a web